+++ internal/cli/serve/command_test.go
@@ -3,5 +3,7 @@ package serve_test
 import "testing"
 
 func TestNothing(t *testing.T) {
+	t.Parallel()
+
 	t.Skip("tests for this package have not been implemented yet")
 }