+++ internal/env/env_test.go
@@ -8,12 +8,15 @@ import (
 )
 
 func TestConstants(t *testing.T) {
+	t.Parallel()
+
 	assert.Equal(t, "LISTEN_ADDR", string(ListenAddr))
 	assert.Equal(t, "LISTEN_PORT", string(ListenPort))
 	assert.Equal(t, "TEMPLATE_NAME", string(TemplateName))
 	assert.Equal(t, "CONFIG_FILE", string(ConfigFilePath))
 	assert.Equal(t, "DEFAULT_ERROR_PAGE", string(DefaultErrorPage))
 	assert.Equal(t, "DEFAULT_HTTP_CODE", string(DefaultHTTPCode))
+	assert.Equal(t, "SHOW_DETAILS", string(ShowDetails))
 }
 
 func TestEnvVariable_Lookup(t *testing.T) {
@@ -26,6 +29,7 @@ func TestEnvVariable_Lookup(t *testing.T) {
 		{giveEnv: ConfigFilePath},
 		{giveEnv: DefaultErrorPage},
 		{giveEnv: DefaultHTTPCode},
+		{giveEnv: ShowDetails},
 	}
 
 	for _, tt := range cases {