+++ internal/http/server_test.go
@@ -3,5 +3,7 @@ package http
 import "testing"
 
 func TestNothing(t *testing.T) {
+	t.Parallel()
+
 	t.Skip("tests for this package have not been implemented yet")
 }