+++ internal/version/version_test.go
@@ -5,6 +5,8 @@ import (
 )
 
 func TestVersion(t *testing.T) {
+	t.Parallel()
+
 	for give, want := range map[string]string{
 		// without changes
 		"vvv":     "vvv",